moved OSHEntities to osh package
