+++ oshdb/src/main/java/org/heigit/bigspatialdata/oshdb/osh/OSHEntity.java
@@ -14,7 +14,6 @@ import java.util.function.Predicate;
 import javax.annotation.Nonnull;
 import org.heigit.bigspatialdata.oshdb.osm.OSMEntity;
 import org.heigit.bigspatialdata.oshdb.osm.OSMType;
-import org.heigit.bigspatialdata.oshdb.util.OSHEntities;
 import org.heigit.bigspatialdata.oshdb.util.OSHDBBoundingBox;
 import org.heigit.bigspatialdata.oshdb.util.OSHDBTagKey;
 import org.heigit.bigspatialdata.oshdb.util.OSHDBTimestamp;
